Move 'Range' object to its own file
